Refactor code related to Project CRUD  - Move shared code to common.js;  - Delete current.js
