+++ src/components/Sidebar.js
@@ -11,22 +11,18 @@ const sidebar = () => {
     'font-weight-bold',
   );
 
-  // Projects
   sidebarHeading.textContent = 'Projects';
   UISidebar.appendChild(sidebarHeading);
 
-  // Add project button
   const btnAddProject = document.createElement('button');
   btnAddProject.id = 'btn-add-project';
   btnAddProject.classList.add('btn', 'btn-success', 'float-right');
   btnAddProject.innerHTML = '<i class="fas fa-plus mr-2"></i>Add Project';
 
-  // Set attributes
   btnAddProject.setAttribute('data-toggle', 'modal');
   btnAddProject.setAttribute('data-target', '#projectModal');
 
   sidebarHeading.append(btnAddProject);
-  // Projects container
   const UIProjectsContainer = document.createElement('div');
   UIProjectsContainer.id = 'projects-container';
   UISidebar.append(UIProjectsContainer);
