+++ src/event_listners/taskEventListners.js
@@ -10,9 +10,7 @@ import {
 
 const { DateTime } = require('luxon');
 
-// Task event Listners
 
-// Add new project on modal save button click
 const addNewTask = () => {
   const btnAddNewTask = document.querySelector('#add-task');
 
@@ -26,9 +24,7 @@ const addNewTask = () => {
 
     const now = DateTime.local();
     let task = new Task(name, desc, date, priority, now);
-    // Inser new task modal
     const taskModal = document.querySelector('#taskModal');
-    // Validate task
     if (!task.validate()) {
       return false;
     }
@@ -37,15 +33,12 @@ const addNewTask = () => {
     task = db.saveTask(task);
     renderNewTask(task);
 
-    // Reset add task modal
     resetAddTaskModal();
-    // Dismiss the modal
     taskModal.querySelector('[data-dismiss="modal"]').click();
     return true;
   });
 };
 
-// Open update task modal
 const openUpdateTaskModal = () => {
   const btnProjectDetails = document.querySelector('.project-tasks-row');
   btnProjectDetails.addEventListener('click', (e) => {
@@ -70,7 +63,6 @@ const openUpdateTaskModal = () => {
   });
 };
 
-// Render updated task upon click of btnUpdateTask
 const updateTask = () => {
   const btnUpdateTask = document.querySelector('#update-task');
   btnUpdateTask.addEventListener('click', (e) => {
@@ -84,16 +76,13 @@ const updateTask = () => {
     const date = document.querySelector('#update-task-date').value;
     const priority = document.querySelector('#update-task-priority').value;
 
-    // Update task modal
     const updateTaskModal = document.querySelector('#updateTaskModal');
-    // Validate task
     const now = DateTime.local();
     const task = new Task(name, description, date, priority, now);
 
     if (!task.validate()) {
       return false;
     }
-    // Retrieve the createdAt field of the current task
     const db = new Db();
     const currentTask = db.getCurrentTask(taskId);
     const taskCard = document.querySelector(`.card-${taskId}`);
@@ -115,28 +104,22 @@ const updateTask = () => {
       currentTask.createdAt,
     );
 
-    // Reset update task modal
     resetUpdateTaskModal();
-    // Dismiss the modal
     updateTaskModal.querySelector('[data-dismiss="modal"]').click();
     return true;
   });
 };
 
-// Delete Task
 const deleteTask = () => {
   const btnProjectDetails = document.querySelector('.project-tasks-row');
 
   btnProjectDetails.addEventListener('click', (e) => {
     if (e.target.classList.contains('btn-delete-task')) {
       const taskId = e.target.parentElement.id;
-      // Update UI to reflect deleted task
       removeDeletedTask(taskId);
 
-      // Get project from the datastore
       const db = new Db();
 
-      // Remove deleted project from the datastore (LS)
       db.deleteTask(taskId);
     }
   });
